backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
SQLAlchemy==1.4.41
databases[sqlite]==0.6.2
pydantic==1.10.11
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.1.0
requests==2.31.0
typing_extensions==4.7.1
alembic
psycopg2-binary
